Do not migrate `BlockBody`
